moved emojis to separate component
